requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
Pillow==10.0.0
pycparser==2.21
PyMySQL==1.1.0
sqlparse==0.4.4
uWSGI==2.0.22
